+++ src/dpr/components/_inputs/date-range/utils.ts
@@ -5,16 +5,18 @@ import isBetween from 'dayjs/plugin/isBetween'
 import { components } from '../../../types/api'
 import { DateFilterValue, DateRange, FilterValue } from '../../_filters/types'
 import StartEndDateUtils from '../start-end-date/utils'
-import RelativeDateRange from './types'
+import RelativeDateRange, { RelativeOption } from './types'
 
-const dateIsInBounds = (startDate: dayjs.Dayjs, endDate: dayjs.Dayjs, min: string, max: string) => {
+const dateIsInBounds = (startDate: dayjs.Dayjs | string, endDate: dayjs.Dayjs | string, min: string, max: string) => {
   dayjs.extend(isBetween)
 
   const minDate = dayjs(min)
   const maxDate = dayjs(max)
 
-  const startDateIsBetweenMinAndMax = startDate.isBetween(minDate, maxDate, 'day', '[]')
-  const endDateIsBetweenMinAndMax = endDate.isBetween(minDate, maxDate, 'day', '[]')
+  const startDateIsBetweenMinAndMax = startDate
+    ? (<dayjs.Dayjs>startDate).isBetween(minDate, maxDate, 'day', '[]')
+    : true
+  const endDateIsBetweenMinAndMax = endDate ? (<dayjs.Dayjs>endDate).isBetween(minDate, maxDate, 'day', '[]') : true
 
   return startDateIsBetweenMinAndMax && endDateIsBetweenMinAndMax
 }
@@ -24,6 +26,10 @@ const calcDates = (durationValue: string) => {
   let startDate
 
   switch (durationValue) {
+    case 'none':
+      endDate = ''
+      startDate = ''
+      break
     case 'yesterday':
       endDate = dayjs()
       startDate = endDate.subtract(1, 'day')
@@ -75,8 +81,8 @@ const setValueFromRequest = (filter: FilterValue, req: Request, prefix: string) 
 const getRelativeDateOptions = (min: string, max: string) => {
   if (!min) min = '1977-05-25'
   if (!max) max = '9999-01-01'
-  let options: { value: string; text: string; disabled?: boolean }[] = getRelativeValues()
-  options.forEach((option: { value: string; text: string; disabled?: boolean }) => {
+  let options: RelativeOption[] = getRelativeValues()
+  options.forEach((option: RelativeOption) => {
     if (option.value) {
       const { endDate, startDate } = calcDates(option.value)
       if (!dateIsInBounds(startDate, endDate, min, max)) {
@@ -86,9 +92,11 @@ const getRelativeDateOptions = (min: string, max: string) => {
   })
 
   if (
-    options.every((opt: { value: string; text: string; disabled?: boolean }) => {
-      return opt.value === null || opt.disabled
-    })
+    options
+      .filter((opt: RelativeOption) => opt.value !== 'none')
+      .every((opt: RelativeOption) => {
+        return opt.value === null || opt.disabled
+      })
   ) {
     options = []
   }
@@ -108,13 +116,13 @@ const getRelativeValues = (): {
   disabled?: boolean
 }[] => {
   return [
-    { value: null, text: 'None' },
+    { value: 'none', text: 'None' },
     { value: 'yesterday', text: 'Yesterday' },
     { value: 'tomorrow', text: 'Tomorrow' },
     { value: 'last-week', text: 'Last week' },
     { value: 'next-week', text: 'Next week' },
     { value: 'last-month', text: 'Last month' },
-    { value: 'next-month', text: 'Next-month' },
+    { value: 'next-month', text: 'Next month' },
   ]
 }